mcp-service/requirements.txt
fastmcp==0.1.0
fastapi>=0.104.0
psycopg2-binary==2.9.9
pydantic>=2.8.0
python-dotenv==1.0.0
uvicorn==0.25.0